+++ frontend/apps/coin-flipper/vite.config.js
@@ -1,6 +1,8 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 
+const PROJECT_NAME = require('./package.json').name.split('/').pop();
+
 // https://vitejs.dev/config/
 export default defineConfig(({ command }) => {
     const config = {
@@ -8,9 +10,9 @@ export default defineConfig(({ command }) => {
         base: '/',
     };
 
-    // if (command !== 'serve') {
-    //     config.base = '/prj--reactjs-component-lib/';
-    // }
+    if (command !== 'serve') {
+        config.base = `/prj--reactjs-component-lib/${PROJECT_NAME}`;
+    }
 
     return config;
 });