+++ assets/media/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.1](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/media-assets@0.4.0...@prj--reactjs-component-lib/media-assets@0.4.1) (2023-10-31)
+
+**Note:** Version bump only for package @prj--reactjs-component-lib/media-assets
+
+# [0.4.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/media-assets@0.3.1...@prj--reactjs-component-lib/media-assets@0.4.0) (2023-10-30)
+
+### Features
+
+-   formats ([d81e33d](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/d81e33db14d484d87c5c29b249d9d2cbd3a13560))
+
+## [0.3.1](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/media-assets@0.3.0...@prj--reactjs-component-lib/media-assets@0.3.1) (2023-10-30)
+
+**Note:** Version bump only for package @prj--reactjs-component-lib/media-assets
+
 # [0.3.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/media-assets@0.2.0...@prj--reactjs-component-lib/media-assets@0.3.0) (2023-10-30)
 
 ### Features