+++ frontend/apps/_boilerplate-app/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/boilerplate-app@0.4.0...@prj--reactjs-component-lib/boilerplate-app@0.5.0) (2023-10-31)
+
+### Features
+
+-   **base path:** fixed ([9436eca](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/9436ecafd5addb266153737a53f95922733b9a63))
+
+# [0.4.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/boilerplate-app@0.3.0...@prj--reactjs-component-lib/boilerplate-app@0.4.0) (2023-10-30)
+
+### Features
+
+-   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/f9780c2896d185c8adf83f5af0782939e799b430))
+-   **package:** added package scripts to apps ([15349d0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/15349d0e3d3eac4222a99a42b28d4d67b764557f))
+
 # [0.3.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/boilerplate-app@0.2.1...@prj--reactjs-component-lib/boilerplate-app@0.3.0) (2023-10-28)
 
 ### Features