+++ frontend/apps/counter-simple-app/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/counter-simple-app@0.3.0...@prj--reactjs-component-lib/counter-simple-app@0.4.0) (2023-10-31)
+
+### Bug Fixes
+
+-   **naming issue:** naming ([5082c59](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/5082c599e93805b29d3df21e594e67b83160f658))
+
+### Features
+
+-   **base path:** fixed ([9436eca](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/9436ecafd5addb266153737a53f95922733b9a63))
+-   **naming issue:** namin ([f2bb365](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/f2bb365256593d698336ba08b8b826545f8d6dc4))
+-   small ([ed67551](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/ed6755124ed3073708ca574f7bf1414304d5f15c))
+
+# [0.3.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/counter-simple-app@0.2.1...@prj--reactjs-component-lib/counter-simple-app@0.3.0) (2023-10-30)
+
+### Features
+
+-   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/f9780c2896d185c8adf83f5af0782939e799b430))
+-   **package:** added package scripts to apps ([15349d0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/15349d0e3d3eac4222a99a42b28d4d67b764557f))
+
 ## [0.2.1](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/counter-simple-app@0.2.0...@prj--reactjs-component-lib/counter-simple-app@0.2.1) (2023-10-28)
 
 **Note:** Version bump only for package @prj--reactjs-component-lib/counter-simple-app