+++ frontend/design-systems/storybook/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/storybook@0.3.0...@prj--reactjs-component-lib/storybook@0.4.0) (2023-10-31)
+
+### Features
+
+-   **storybok:** package ([9c6a666](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/9c6a666f8096beff0ede671c56e272a1061baa26))
+
+# [0.3.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/storybook@0.2.0...@prj--reactjs-component-lib/storybook@0.3.0) (2023-10-30)
+
+### Features
+
+-   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/f9780c2896d185c8adf83f5af0782939e799b430))
+
 # 0.2.0 (2023-10-26)
 
 ### Features